cassandra_replication/requirements.txt
cassandra-driver==3.29.3